First attempt at some history keeping. Decided to keep the list in the article, since everything I want to keep records about has to do with articles. I don't really care to record whether people changed their profiles or changed community settings. What I want the history to show mainly is whether people added / changed / removed stories and annotations. This information will be really important when the GUI is made, it is less important now.
Also made a first messy attempt at a "all members" page and pages for each member. I've put the history sections there and in the article page, but that is just for testing. Really only curators should see the article histories and only guides should see member histories. Not sure the history system is as complete as it should be, will reassess next week.