Decided to remove the "anonymous" entry type as a separate type from personifications, because it was just cluttering things up to no useful purpose. Also renamed "characters" because the p-word was just much too long. Now a community can have any number of characters, and the choice is more simple: you can attribute a story/comment/whatever to yourself or to a character.
I also decided (after much wrangling) that the "read" page should NOT show that it was you who wrote a story/whatever if you choose a character attribution. I have not decided exactly how people will get to entries they made with character attribution, but it will be hidden and not as obvious as before when anyone looking over their shoulder would be able to see who really said it. All this anonymization stuff is important as it impacts on the nature of storytelling as distinct from the normal flow of conversation.