I'm going to start checking in things here and noting the release numbers I'm giving them on the web site. This is release 1.0.0 beta 1.3. This information is repeated on http://www.rakontu.org/changes.html
Improved process of canceling the joining of a Rakontu so it is less confusing.

Made it possible for managers to remove inactive members (ie, people who left) if they have not contributed anything.

Changed how newly invited member emails are matched with Google account emails to stop
checking for a case-sensitive match. This is to deal with the fact that for Gmail at least,
what is displayed to the user is not what the Google App Engine system tells Rakontu
that person's email address is. The likelihood of two people having the same email provider
and the same email address (except for a case difference), invited to join the same Rakontu
at the same time, is small enough that the inconvenience of getting the case perfect
is not worth doing. (However, managers need to be aware if the situation <i>does</i>
come up of two invitees with nearly the same emails, that they should
invite them one at a time.)

Put a change-your-nickname link on the new-member page, as a sort of subtle message
that calling yourself "Member 23" forever doesn't have to happen.

Added a Manage-Skin page with previews of all built-in skins, to make it easier
to choose a look for each Rakontu. Had been wanting to do this for a <i>long</i> time.

Added more information for managers to see how long it has been since
each member has been active in the Rakontu.