Two small bug fixes.
First, was getting this error on the Rakontu Sandbox:

------------------------------------------
Traceback (most recent call last):

  File "/base/python_runtime/python_lib/versions/1/google/appengine/ext/webapp/__init__.py", line 511, in __call__
    handler.get(*groups)

  File "/base/data/home/apps/rakontu-sandbox/1-0-0-beta-1-6.338777851352660677/utils.py", line 48, in check_login
    func(request)

  File "/base/data/home/apps/rakontu-sandbox/1-0-0-beta-1-6.338777851352660677/visit.py", line 113, in get
    viewOptions = member.getViewOptionsForLocation("home")

  File "/base/data/home/apps/rakontu-sandbox/1-0-0-beta-1-6.338777851352660677/models.py", line 1625, in getViewOptionsForLocation
    self.createViewOptions()

  File "/base/data/home/apps/rakontu-sandbox/1-0-0-beta-1-6.338777851352660677/models.py", line 1514, in createViewOptions
    db.delete(viewOptionsNow)

  File "/base/python_runtime/python_lib/versions/1/google/appengine/ext/db/__init__.py", line 1300, in delete
    keys = [_coerce_to_key(v) for v in models]

  File "/base/python_runtime/python_lib/versions/1/google/appengine/ext/db/__init__.py", line 362, in _coerce_to_key
    value = value[0]

IndexError: list index out of range 
------------------------------------------

I have no idea why this should happen (suddenly - it never happened before), as there seems to be nothing wrong with my code. But as I've had my head out of the code for a while I don't have the time to get back inside the onion to figure it out. So I've just commented out the deleting code. That means that extra viewOptions objects will pile up, but they are small and this method should not be called often anyway. (I think)

The second bug was just a typo where I was giving the Rakontu's id instead of its pointer to a method.